Moved UserInfo type to store package, to avoid circular refs.
