+++ api/structs.go
@@ -54,14 +54,6 @@ type SSHKey struct {
 	PublicKey string `json:"public_key"`
 }
 
-// UserInfo is the response structure of a GetUserInfo operation
-type UserInfo struct {
-	Sub       string `json:"sub"`
-	FirstName string `json:"firstname"`
-	Surname   string `json:"surname"`
-	Email     string `json:"email"`
-}
-
 // LogLevel
 type LogLevel uint8
